[AI-Flow]Rename notification_service.master to notification_service.server
