+++ lib/notification_service/tests/test_mongo_notification.py
@@ -22,7 +22,7 @@ from typing import List
 from notification_service.base_notification import BaseEvent, EventWatcher
 from notification_service.client import NotificationClient
 from notification_service.mongo_event_storage import MongoEventStorage
-from notification_service.master import NotificationServer
+from notification_service.server import NotificationServer
 from notification_service.service import NotificationService
 
 properties = {'enable.idempotence': 'True'}
